+++ src/index.js
@@ -3,8 +3,3 @@ import Api from './modules/Api.js';
 
 const ApiCalls = new Api();
 ApiCalls.GetMeals();
-
-const saveComment = (data) => {
-  ApiCalls.AddComment(data);
-};
-export default saveComment;
